refactor: put session code to its own file
